Update tests.
The Buffer API changed, the test wasn't changed to accommodate this
fact.  The expected results for rendering the menu were not up to date
with reality anymore.

Statically linking against glibc makes Valgrind display lots of errors,
that do not occur when linking dynamically.  For this reason static
linking is disabled again.